feat: change datareducer to dataSlice
